requirements.txt
django>=2.2,<2.3
djangorestframework==3.9.1
django-filter>=2,<2.1
